Consolidate index and runIntegrationTests files
